backend/requirements.txt
Flask
python-dotenv
flask-cors
requests
google-generativeai